requirements.txt
astroid==1.5.3
autopep8==1.3.2
awscli==1.11.141
backports.functools-lru-cache==1.4
boto==2.48.0
boto3==1.4.6
botocore==1.6.8
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
configparser==3.5.0
cookies==2.2.1
dicttoxml==1.7.4
docutils==0.14
enum34==1.1.6
extras==1.0.0
fixtures==3.0.0
funcsigs==1.0.2
futures==3.1.1
fuzzywuzzy==0.15.1
idna==2.6
isort==4.2.15
Jinja2==2.9.6
jmespath==0.9.3
lazy-object-proxy==1.3.1
lettuce==0.2.23
linecache2==1.0.0
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
moto==1.1.1
pbr==3.1.1
py==1.4.34
pyaml==17.8.0
pyasn1==0.3.3
pycodestyle==2.3.1
pylint==1.7.2
pytest==3.2.1
python-dateutil==2.6.1
python-Levenshtein==0.12.0
python-mimeparse==1.6.0
python-subunit==1.2.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.10
singledispatch==3.4.0.3
six==1.10.0
sure==1.4.6
testtools==2.3.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
xmltodict==0.11.0
